Allow adding render engine after registering geometry (#11820)
* Allows geometries to be added before renderers

This removes the old requirement that all renderers needed to be
added before any geometry. Now, geometry that has been registered
and has had the perception property assigned will automatically
be added to any subsequent render engine that gets added.

* Eliminate the two-way coupling between RenderEngine and InternalGeometry

InternalGeometry used to have a list of all render engines it was
registered with. This required updating *two* things every time a geometry
was added/removed from a render engine.

This modifies RenderEngine to report if an id has been registered with it.
Now, rather than querying the internal geometry if it is a render engine
(and hope that it is correct), we simply ask the engine directly.